chore: Update dependencies and styles
This commit updates the package.json file to include the latest versions of the `primeflex` and `primeng` libraries. It also adds the necessary imports for the `DashComponent` in the `app-routing.module.ts` file. Additionally, the `styles.css` file is modified to include the new theme and CSS files from the `primeng` and `primeflex` libraries. Finally, the `app.module.ts` file is updated to include the necessary modules for the `DashComponent` and the `PrimeNGModule`.